Travis-CI: use the default build system. Use a double for division calculations per SonarCloud recommendations in PlacardRenderer. Use @Nullable annotation to try and make SonarCloud less noisey about the Forge Capability API instance injection.

+++ src/main/java/net/aeronica/mods/mxtune/render/PlacardRenderer.java
@@ -40,7 +40,7 @@ public class PlacardRenderer
 
     private final ResourceLocation placardTextures = new ResourceLocation(MXTuneMain.prependModID("textures/gui/status_widgets.png"));
     private final float PLACARD_RANGE = 32.0f;
-    private int index = 0;
+    private double index = 0;
 
     private PlacardRenderer() {}
